Downgrade airbrake from 11.0.3 to 7.4.0
Two of the custom information texts tests fail sometimes because the
page takes too long to load. After upgrading airbrake 9ac8528cb, maybe
because of the way requests are intercepted, or maybe because of the
rbtree3 dependency, it looks like requests are a bit slower and so these
tests were failing 100% of the time.

While we investigate the issue, we're downgrading to the latest version
of Airbrake where the tests still pass.