Add consistency between deploy-secrets and installer
The installer creates the application in the following path:
home_dir: "/home/{{{ deploy_user }}"

We update the variable "deploy_to" in the deploy-secrets
file to be consistent with the installer.

While we are modifying the file we also update the "user"
variable in staging and preproduction sections with the
installer default value: "deploy_user: deploy" the same as
we already have in the production section.