Enable L.Control.Pan for all application maps
We only need the new controls for the mobile
version as it is impossible to move map panes after
disabling `dragging` and `tap` options. In contrast,
desktop users can drag the map by clicking and
dragging the mouse. However, I finally decided to
show the pan controls for desktop and mobile
devices to keep consistency and considering that
some users can browse the website through both
kind on devices.